Add lab 3
